refactor: comment 패키지 밖으로 이동
