+++ src/main/java/dynamicquad/agilehub/comment/domain/Comment.java
@@ -1,4 +1,4 @@
-package dynamicquad.agilehub.issue.comment.domain;
+package dynamicquad.agilehub.comment.domain;
 
 import dynamicquad.agilehub.global.domain.BaseEntity;
 import dynamicquad.agilehub.issue.domain.Issue;
